# docs/openapi.yaml
openapi: 3.0.3
info:
  title: WattPilot API
  version: 1.0.0
  description: |
    WattPilot V1 REST API specification.
    Optimizes EV charging against Norwegian hourly electricity prices.
    V1 uses manually registered EVs, continuous charging windows, and mock charging;
    no vehicle-manufacturer control is performed.
servers:
  - url: http://localhost:8081/api/v1
    description: Local development
...
paths:
  /auth/signup:
    post:
      tags: [Auth]
      summary: Create an account
      operationId: signUp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignUpRequest' }
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out and revoke refresh token
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '204': { description: Logged out }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request syntax or parameters
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Conflict:
      description: Request conflicts with current resource state
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    PriceArea:
      type: string
      enum: [NO1, NO2, NO3, NO4, NO5]
    UserStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
...
    SignUpRequest:
      type: object
      required: [email, password, name, defaultPriceArea]
      properties:
        email: { type: string, format: email, maxLength: 255, example: iris@example.com }
        password: { type: string, format: password, minLength: 8, maxLength: 72, writeOnly: true, example: WattPilot123! }
        name: { type: string, minLength: 1, maxLength: 100, example: Iris }
        defaultPriceArea: { $ref: '#/components/schemas/PriceArea' }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: iris@example.com }
        password: { type: string, format: password, writeOnly: true, example: WattPilot123! }
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string, minLength: 1, writeOnly: true }
    TokenResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn]
      properties:
        accessToken: { type: string, readOnly: true }
        refreshToken: { type: string, readOnly: true }
        tokenType: { type: string, enum: [Bearer], example: Bearer }
        expiresIn: { type: integer, format: int64, description: Access-token lifetime in seconds, example: 3600 }
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/TokenResponse'
        - type: object
          required: [user]
          properties:
            user: { $ref: '#/components/schemas/User' }
    User:
      type: object
      required: [id, email, name, defaultPriceArea, status, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64, minimum: 1, example: 1 }
        email: { type: string, format: email }
        name: { type: string }
        defaultPriceArea: { $ref: '#/components/schemas/PriceArea' }
        status: { $ref: '#/components/schemas/UserStatus' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    FieldError:
      type: object
      required: [field, reason]
      properties:
        field: { type: string, example: targetBatteryPercent }
        rejectedValue: { nullable: true, example: 120 }
        reason: { type: string, example: must be less than or equal to 100 }
    ErrorResponse:
      type: object
      required: [timestamp, status, code, message, path, traceId]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 400 }
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: Request validation failed. }
        path: { type: string, example: /api/v1/charging-plans }
        traceId: { type: string, example: 7b9f52e8a61b4a74 }
        fieldErrors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }